create events sub-reducer
